[cpu] move run_until_alarm out of crate cpu into crate cli.
The motivation is that this puts the "main loop" in the CLI itself,
meaning that the library doesn't implement the main control flow.
This should make it easier to integrate the simulator as a library.
